b2z-grid-utilization: the pairformer block runs at 97.3 % core utilization, so narrow grids are 0.30 s of a 6.6 s deficit

+++ tt_bio/kernels/triatt_sdpa/compute/compute_common.hpp
@@ -54,7 +54,7 @@ void max_block_inplace(uint32_t in0, uint32_t in1) {
         acquire_dst();
         copy_tile(in0, i, dst_reg_0);
         copy_tile(in1, i, dst_reg_1);
-        binary_max_tile(dst_reg_0, dst_reg_1, dst_reg_0, static_cast<int>(VectorMode::C));
+        binary_max_tile(dst_reg_0, dst_reg_1, dst_reg_0, VectorMode::C);
         pack_tile(dst_reg_0, in0);
         release_dst();
     }
@@ -81,7 +81,7 @@ void max_block(uint32_t in0, uint32_t in1, uint32_t out_cb, uint32_t num_tiles) 
         acquire_dst();
         copy_tile(in0, i, dst_reg_0);
         copy_tile(in1, i, dst_reg_1);
-        binary_max_tile(dst_reg_0, dst_reg_1, dst_reg_0, static_cast<int>(VectorMode::C));
+        binary_max_tile(dst_reg_0, dst_reg_1, dst_reg_0, VectorMode::C);
         pack_tile(dst_reg_0, out_cb, i);
         release_dst();
     }
